Moved HelpDialogComponent into its own file and subdirectory and moved TrackPlayerControlsComponent and TrackPlayerProgressComponent to their own subdirectories
