+++ src/app/components/_index.scss
@@ -1,6 +1,6 @@
 @forward 'album-display/album-display.component.theme';
 @forward 'app/app.component.theme';
 @forward 'devices/devices.component.theme';
-@forward 'settings-menu/help-dialog.component.theme';
+@forward 'settings-menu/help-dialog/help-dialog.component.theme';
 @forward 'settings-menu/settings-menu.component.theme';
-@forward 'track-player/track-player.component.theme';
+@forward 'track-player/track-player-controls/track-player-controls.component.theme';
