+++ src/app/components/settings-menu/help-dialog/_help-dialog.component.theme.scss
@@ -1,4 +1,4 @@
-@use '@angular/material' as mat;
+@use '@angular/material/index' as mat;
 @use 'sass:map';
 
 @mixin help-dialog-theme($theme) {
