+++ src/app/components/track-player/track-player-controls/_track-player-controls.component.theme.scss
@@ -1,4 +1,4 @@
-@use '@angular/material' as mat;
+@use '@angular/material/index' as mat;
 @use 'sass:map';
 
 @mixin track-player-theme($theme) {
