+++ src/app/components/track-player/track-player-controls/track-player-controls.component.css
@@ -1,0 +1,14 @@
+.volume-slider {
+  width: 100%;
+  max-width: 40%;
+}
+
+.volume-button {
+  height: 32px !important;
+  width: 32px !important;
+  font-size: 32px !important;
+  line-height: 32px !important;
+  vertical-align: center !important;
+}
+
+
