+++ src/app/components/track-player/track-player-controls/track-player-controls.component.ts
@@ -2,6 +2,7 @@ import { Component, ElementRef, Input, OnDestroy, OnInit } from '@angular/core';
 import { MatSliderChange } from '@angular/material/slider';
 import { Select, Store } from '@ngxs/store';
 import { Observable, Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 import {
   ChangeDeviceVolume,
   ChangeRepeatState,
@@ -10,12 +11,12 @@ import {
   ToggleLiked,
   TogglePlaying,
   ToggleShuffle
-} from '../../core/playback/playback.actions';
-import { PREVIOUS_VOLUME } from '../../core/playback/playback.state';
-import { PlayerControlsOptions } from '../../core/settings/settings.model';
-import { SettingsState } from '../../core/settings/settings.state';
-import { InactivityService } from '../../services/inactivity/inactivity.service';
-import { StorageService } from '../../services/storage/storage.service';
+} from '../../../core/playback/playback.actions';
+import { PREVIOUS_VOLUME } from '../../../core/playback/playback.state';
+import { PlayerControlsOptions } from '../../../core/settings/settings.model';
+import { SettingsState } from '../../../core/settings/settings.state';
+import { InactivityService } from '../../../services/inactivity/inactivity.service';
+import { StorageService } from '../../../services/storage/storage.service';
 
 // Default values
 const DEFAULT_VOLUME = 50;
@@ -32,6 +33,9 @@ const VOLUME_HIGH_ICON = 'volume_up';
 const VOLUME_LOW_ICON = 'volume_down';
 const VOLUME_MUTE_ICON = 'volume_off';
 
+const ICON_CLASS_PRIMARY = 'track-player-icon';
+const ICON_CLASS_ACCENT = 'track-player-icon-accent';
+
 // Keys
 const REPEAT_OFF = 'off';
 const REPEAT_CONTEXT = 'context';
@@ -40,7 +44,7 @@ const REPEAT_TRACK = 'track';
 @Component({
   selector: 'app-track-player-controls',
   templateUrl: './track-player-controls.component.html',
-  styleUrls: ['./track-player.component.css']
+  styleUrls: ['./track-player-controls.component.css']
 })
 export class TrackPlayerControlsComponent implements OnInit, OnDestroy {
   private ngUnsubscribe = new Subject();
@@ -61,20 +65,24 @@ export class TrackPlayerControlsComponent implements OnInit, OnDestroy {
               private element: ElementRef) {}
 
   ngOnInit(): void {
-    this.showPlayerControls$.subscribe((option) => {
-      const isFading = option === PlayerControlsOptions.Fade;
-      // Make sure we display player controls if previously off
-      if (this.fadePlayerControls && !isFading) {
-        this.fadeControls(false);
-      }
-      this.fadePlayerControls = isFading;
-    });
+    this.showPlayerControls$
+      .pipe(takeUntil(this.ngUnsubscribe))
+      .subscribe((option) => {
+        const isFading = option === PlayerControlsOptions.Fade;
+        // Make sure we display player controls if previously off
+        if (this.fadePlayerControls && !isFading) {
+          this.fadeControls(false);
+        }
+        this.fadePlayerControls = isFading;
+      });
 
-    this.inactivity.inactive$.subscribe((isInactive) => {
-      if (this.fadePlayerControls) {
-        this.fadeControls(isInactive);
-      }
-    });
+    this.inactivity.inactive$
+      .pipe(takeUntil(this.ngUnsubscribe))
+      .subscribe((isInactive) => {
+        if (this.fadePlayerControls) {
+          this.fadeControls(isInactive);
+        }
+      });
   }
 
   ngOnDestroy(): void {
@@ -147,10 +155,11 @@ export class TrackPlayerControlsComponent implements OnInit, OnDestroy {
   }
 
   getRepeatClass(repeatState: string): string {
-    if (repeatState === REPEAT_OFF) {
-      return 'track-player-icon';
+    let repeatClass = ICON_CLASS_PRIMARY;
+    if (repeatState !== REPEAT_OFF) {
+      repeatClass = ICON_CLASS_ACCENT;
     }
-    return 'track-player-icon-accent';
+    return repeatClass;
   }
 
   getVolumeIcon(volume: number): string {
@@ -166,8 +175,9 @@ export class TrackPlayerControlsComponent implements OnInit, OnDestroy {
   private fadeControls(isFaded: boolean): void {
     if (this.element.nativeElement) {
       this.element.nativeElement.animate(
-        {opacity: isFaded ? 0 : 1},
         {
+          opacity: isFaded ? 0 : 1
+        }, {
           duration: FADE_DURATION,
           fill: 'forwards',
           easing: 'ease-out'
